requirements.txt
# DeepAgents 虚拟环境依赖
# 安装命令: pip install -r requirements.txt

aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
attrs==25.4.0
bracex==2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
deepagents==0.4.3
distro==1.9.0
dockerfile-parse==2.0.1
docstring_parser==0.17.0
e2b==2.14.0
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.64.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.10
langchain-anthropic==1.3.3
langchain-core==1.2.14
langchain-deepseek==1.0.1
langchain-google-genai==4.2.1
langchain-mcp-adapters==0.2.1
langchain-openai==1.1.10
langchain-tavily==0.2.17
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.6
lxml==6.0.2
markdown-it-py==4.0.0
mcp==1.26.0
mdurl==0.1.2
multidict==6.7.1
openai==2.21.0
orjson==3.11.7
ormsgpack==1.12.2
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049096285/work
pillow==12.1.1
propcache==0.4.1
protobuf==7.34.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-pptx==1.0.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
reportlab==4.4.10
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.2.0
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
wcmatch==10.1
websockets==15.0.1
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
